% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getDataAndShowStatus}
\alias{getDataAndShowStatus}
\title{getDataAndShowStatus}
\usage{
getDataAndShowStatus(responseObject)
}
\arguments{
\item{responseObject}{objet de reponse HTTP httr}
}
\description{
Recupere les status et les informations presentes dans les entetes de reponse HTTP
 ainsi que dans la partie metadata de la reponse
}
\keyword{internal}
