% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_var_env.R
\name{getDataVarPretty}
\alias{getDataVarPretty}
\title{getDataVarPretty}
\usage{
getDataVarPretty(nameVar, methodVar = NULL, varPretty, token)
}
\arguments{
\item{nameVar}{(string) name of the variable to plot}

\item{methodVar}{(string) name of the method used to collect data}

\item{varPretty}{(data.frame) from getVarPretty}

\item{token}{(string) a token from getToken function}
}
\value{
WSResponse
}
\description{
getDataVarPretty
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
 aToken <- getToken("guest@phis.fr","guest")
 token <- aToken$data
 varPrettyTot <- getVarPretty(token = token)
 getDataVarPretty(nameVar = "temperature", varPretty = varPrettyTot, token = token)
}
}
