% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getEnvironment}
\alias{getEnvironment}
\title{retrieves the environmental mesures of an experiment from the web service}
\usage{
getEnvironment(token, variableCategory = "", startDate = "",
  endDate = "", variables = "", facility = "", experimentURI = "",
  page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{variableCategory}{character, a category of variables}

\item{startDate}{data > startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{endDate}{data < startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{variables}{list of variables for the request (Ex : "wind speed_weather station_meter per second")}

\item{facility}{place of the experiment (Ex : "http://www.phenome-fppn.fr/m3p/ec3")}

\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves environmental mesures of an experiment or by dates
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 getEnvironment(aToken$data,page=3,pageSize=100,startDate="2012-02-21",endDate = "2012-03-21")
 test<-getEnvironment(aToken$data,
       experimentURI="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 test$data
 getEnvironment(aToken$data,experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
 startDate="2012-02-21",endDate="2012-02-15 19:20:30")
 getEnvironment(aToken$data,experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
    facility="http://www.phenome-fppn.fr/m3p/ec3",
    variables="wind speed_weather station_meter per second")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
