% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getEnvironmentData}
\alias{getEnvironmentData}
\title{getEnvironmentData}
\usage{
getEnvironmentData(token, variable = "", startDate = "",
  endDate = "", sensor = "", page = NULL, pageSize = NULL,
  dateSortAsc = TRUE, verbose = FALSE)
}
\arguments{
\item{token}{a token from getToken function}

\item{variable}{Search by the uri of a variable}

\item{startDate}{Search from start date}

\item{endDate}{Search to end date}

\item{sensor}{Search by the uri of a sensor}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{dateSortAsc}{boolean, sort date in ascending order if TRUE}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the environmental data from a variable or a sensor
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 vars <- getEnvironmentData(aToken$data,category="imagery",
          variable = "http://www.phenome-fppn.fr/ues/id/variables/v005")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
