% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getExperimentById}
\alias{getExperimentById}
\title{getExperimentById}
\usage{
getExperimentById(token, experimentURI = "", page = NULL,
  pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
retrieves the informations for one experiment
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 publicExp<-getExperimentById(aToken$data,
        experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 publicExp$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
