% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getImagesAnalysis}
\alias{getImagesAnalysis}
\title{retrieves the images analysis data from the web service}
\usage{
getImagesAnalysis(token, experimentURI = "", variablesName = list(),
  labelView = "", provider = "", date = "", page = NULL,
  pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{experimentURI}{URI of the experiment}

\item{variablesName}{list, variable names of images analysis (ex : "objAreaSum")}

\item{labelView}{character, label view of an image}

\item{provider}{character, provider of data}

\item{date}{character, data for one day (format: YYYY-MM-DD)}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves data from image analysis
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 myImages<-getImagesAnalysis(token = aToken$data,
           experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
           variablesName = list("objAreaSum"),pageSize = 100000,verbose=FALSE)
 head(myImages$data)
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
