% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getLabelViewByExperimentById}
\alias{getLabelViewByExperimentById}
\title{getLabelViewByExperimentById}
\usage{
getLabelViewByExperimentById(token, experimentURI = "", viewType = "",
  cameraAngle = "", provider = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{experimentURI}{experiment unique identifier}

\item{viewType}{type of view, top or side}

\item{cameraAngle}{numeric, angle of the camera}

\item{provider}{origin of the data}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves LabelViews used in a specific experiment
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
# Not run (is an internal function)
# aToken = getToken("guestphis@supagro.inra.fr","guestphis")$data
# publicLabelView <- getLabelViewByExperimentById(aToken,
#      experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
# publicLabelView$data
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
