% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getPlantEnvironment}
\alias{getPlantEnvironment}
\title{retrieves the environmental mesures of a plant from the web service}
\usage{
getPlantEnvironment(token, plantURI = "", variableCategory = "",
  startDate = "", endDate = "", variables = "", facility = "",
  experimentURI = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{plantURI}{plant URI}

\item{variableCategory}{character, a category of variables}

\item{startDate}{data > startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{endDate}{data < startDate (Format: YYYY-MM-DD or YYYY-MM-DD HH:MM:SS )}

\item{variables}{list of variables for the request (Ex: "wind speed_weather station_meter per second")}

\item{facility}{place of the experiment (Ex: "http://www.phenome-fppn.fr/m3p/ec3")}

\item{experimentURI}{URI of the experiment}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves environmental mesures of a plant or by dates
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
# not run (is an internal function!!!)
# aToken<-getToken("guestphis@supagro.inra.fr","guestphis")$data
# myplant<-getPlantEnvironment(aToken,
#       plantURI="http://www.phenome-fppn.fr/m3p/arch/2011/c11005809",
#       experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
# myplant$data
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
