% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getPlants}
\alias{getPlants}
\title{retrieves the context of plant linked to an experiment from the web service}
\usage{
getPlants(token, plantAlias = "", experimentURI = "",
  germplasmURI = "", page = NULL, pageSize = NULL, verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{plantAlias}{an alias of plant}

\item{experimentURI}{URI of the experiment}

\item{germplasmURI}{filter by genotype}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves context of plant linked to an experiment
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
 aToken<-getToken("guestphis@supagro.inra.fr","guestphis")$data
 plantes<-getPlants(aToken,experimentURI ="http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
