% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getProjects}
\alias{getProjects}
\title{retrieves the list of projects from the web service}
\usage{
getProjects(token, projectName = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{projectName}{Name of the project to search}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the list of projects in the WS
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_public")
 accesToken = getToken("guestphis@supagro.inra.fr","guestphis")
 getProjects(accesToken$data)
 getProjects(accesToken$data, page = 1)
 getProjects(accesToken$data, page = 3, pageSize = 100)
 getProjects(accesToken$data, projectName = "PHIS_Publi")
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
