% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getResponseFromWS}
\alias{getResponseFromWS}
\title{getResponseFromWS}
\usage{
getResponseFromWS(resource, paramPath = NULL, attributes,
  type = "application/json", verbose = FALSE)
}
\arguments{
\item{verbose}{logical FALSE by default, if TRUE display information about the progress}

\item{responseObject}{object HTTP httr}
}
\description{
Create an URL to call the WS and retrun a formatted response of WSResponse class.
}
\keyword{internal}
