% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getResponseFromWS2}
\alias{getResponseFromWS2}
\title{getResponseFromWS2}
\usage{
getResponseFromWS2(resource, paramPath = NULL, attributes, authorization,
  type = "application/json", verbose = FALSE)
}
\arguments{
\item{verbose}{logical FALSE by default, if TRUE display information about the progress}

\item{responseObject}{object HTTP httr}
}
\description{
Create an URL to call the WS2 and retrun a formatted response of WSResponse class.
}
\keyword{internal}
