% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getToken}
\alias{getToken}
\title{retrieves a user identifier for connexion to the web service}
\usage{
getToken(login, password, verbose = FALSE)
}
\arguments{
\item{login}{login of the user to create the token}

\item{password}{password of the user to create the token}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
a session token user identifier in the WS
}
\description{
Retrieves a user identifier for connexion to the WebService (WS)
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
aToken <- getToken("guestphis@supagro.inra.fr","guestphis")
aToken$data
}
}
