% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getTokenResponseWS2}
\alias{getTokenResponseWS2}
\title{getTokenResponseWS2}
\usage{
getTokenResponseWS2(resource, attributes, type = "application/json",
  verbose = FALSE)
}
\arguments{
\item{resource}{character, an resource from the web service api}

\item{attributes}{a list containing a login and a password}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
responseObject an object HTTP httr
}
\description{
Create a token to call the webservice for authentication and
returns a formatted response of WSResponse class.
}
\details{
This function is OK for the second version of the web service
 (a POST call with an invisible request using a correct JSON list in a body)
}
\seealso{
https://brapi.docs.apiary.io/#introduction/structure-of-the-response-object
}
\keyword{internal}
