% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getVariables2}
\alias{getVariables2}
\title{Retrieves the data of a service from the WS2}
\usage{
getVariables2(token, uri = "", label = "", trait = "", method = "",
  unit = "", pageSize = NULL, page = NULL, verbose = FALSE)
}
\arguments{
\item{token}{a token from getToken function}

\item{uri}{Search by the uri of an experiment}

\item{label}{Search by label}

\item{trait}{Search by trait uri}

\item{method}{Search by method uri}

\item{unit}{Search variables by unit uri}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}

\item{ressource}{the name of the service to call}

\item{paramPath}{the extension of the service to call, default to NULL}

\item{attributes}{The list of attributes to give to the GET request}

\item{type}{The type of the output, default to application/json}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Create an URL to call the WS and retrun a formatted response of WSResponse class.

Retrieves the variable descriptions, trait, method and unit covered by the variable
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 vars <- getVariablesByCategory(aToken$data,category="imagery",
          experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
\keyword{internal}
