% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getVariablesByCategory}
\alias{getVariablesByCategory}
\title{getVariablesByCategory}
\usage{
getVariablesByCategory(token, category = "", experimentURI = "",
  imageryProvider = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{category}{Name of the category to search}

\item{experimentURI}{URI of the experiment}

\item{imageryProvider}{character, provider of the images}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves the variable by categories (environment or setpoint...)
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_public")
 aToken = getToken("guestphis@supagro.inra.fr","guestphis")
 vars <- getVariablesByCategory(aToken$data,category="imagery",
          experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01")
 vars$data
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
