% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{getWatering}
\alias{getWatering}
\title{retrieves the irrigation data from the web service}
\usage{
getWatering(token, experimentURI = "", variablesName = list(),
  provider = "", date = "", page = NULL, pageSize = NULL,
  verbose = FALSE)
}
\arguments{
\item{token}{a token}

\item{experimentURI}{URI of the experiment}

\item{variablesName}{list, variable names of images analysis}

\item{provider}{character, provider of data}

\item{date}{character, data for one day (format: YYYY-MM-DD)}

\item{page}{displayed page (pagination Plant Breeding API)}

\item{pageSize}{number of elements by page (pagination Plant Breeding API)}

\item{verbose}{logical FALSE by default, if TRUE display information about the progress}
}
\value{
WSResponse object
}
\description{
Retrieves irrigation data
}
\details{
You have to execute the getToken() function first to have access to the web
service
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_public")
 accesToken = getToken("guestphis@supagro.inra.fr","guestphis")
 mywater<-getWatering(token=accesToken$data,
         experimentURI = "http://www.phenome-fppn.fr/m3p/ARCH2012-01-01",
         variablesName = list("weightBefore"),pageSize=100000,verbose=FALSE)
 head(mywater$data)
}
}
\seealso{
http://docs.brapi.apiary.io/#introduction/url-structure
}
