% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{initializeClientConnection}
\alias{initializeClientConnection}
\title{initializeClientConnection}
\usage{
initializeClientConnection(apiID, url = "")
}
\arguments{
\item{apiID}{character, a character name of an API ("ws_public" or "ws_private")}

\item{url}{character, if apiID is private add the url of the chosen API, containing the IP,
the port and the path of the WS}
}
\description{
load name space and connexion parameters of the webservice.
Execute only once at the beginning of the requests.
In the case of a WebService change of address or a renaming of services, please edit this list.
and execute the function.
}
