% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listVariables.R
\name{listVariables}
\alias{listVariables}
\title{listVariables}
\usage{
listVariables(token)
}
\arguments{
\item{token}{a token}
}
\value{
list
}
\description{
listVariables
}
\examples{
\donttest{
listVariables()
}
}
