% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_encours.R
\name{ObjectType}
\alias{ObjectType}
\title{ObjectType}
\usage{
ObjectType(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
string
}
\description{
Returns the type of object received by R Development function
}
\keyword{internal}
