% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_var_env.R
\name{plot2Vars}
\alias{plot2Vars}
\title{plot2Vars}
\usage{
plot2Vars(nom_var1, nom_var2, startDate = NULL, endDate = NULL, token)
}
\arguments{
\item{nom_var1}{(string) name of the first variable to plot}

\item{nom_var2}{(string) name of the second variable to plot}

\item{startDate}{(date format %Y-%m-%d) date from which to plot}

\item{endDate}{(date format %Y-%m-%d) date to which to plot}

\item{token}{(string) a token from getToken function}
}
\description{
plot2Vars
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
 aToken <- getToken("guest@phis.fr","guest")
 token <- aToken$data
 plot2Vars(nom_var1 = "temperature", nom_var2 = "radiation", token = token)
 plot2Vars(nom_var1 = "temperature", nom_var2 = "radiation", startDate = "2017-06-26", endDate = "2017-06-28", token = token)
}

}
