% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarRel.R
\name{plotVarRel}
\alias{plotVarRel}
\title{Plot two Environmental variables relationship}
\usage{
plotVarRel(varX, varY, startDate = NULL, endDate = NULL,
  trend = FALSE, wsUrl = "www.opensilex.org/openSilexAPI/rest/", token)
}
\arguments{
\item{varX}{name of the variable to plot in X axis}

\item{varY}{name of the variable to plot in Y axis}

\item{startDate}{date from which to plot}

\item{endDate}{date to which to plot}

\item{trend}{logical, draw the trend of the scatterplot}

\item{wsUrl}{url of the webservice}

\item{token}{a token from \code{\link{getToken}} function}
}
\description{
Plot two Environmental variables relationship
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 plotVarRel( "temperature_hourly instant temperature_degree celsius",
       "radiation_hourly global radiation_joule per square centimeter", token = token, trend = TRUE)
}

}
