% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataVarPretty.R
\name{getDataVarPretty}
\alias{getDataVarPretty}
\title{Get Data from WS2 and formate them}
\usage{
getDataVarPretty(nameVar, methodVar = NULL, varPretty, token)
}
\arguments{
\item{nameVar}{name of the variable to plot}

\item{methodVar}{name of the method used to collect data}

\item{varPretty}{from \code{\link{getVarPretty}}}

\item{token}{a token from \code{\link{getToken}} function}
}
\value{
WSResponse
}
\description{
Get Data from WS2 and formate them
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 varPrettyTot <- getVarPretty(token = token)
 getDataVarPretty(nameVar = "temperature", varPretty = varPrettyTot, token = token)
}
}
