% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDF.R
\name{getDF}
\alias{getDF}
\title{getDF}
\usage{
getDF(nameVar, token, smoothing = FALSE, sensor = NULL,
  endDate = NULL, startDate = NULL)
}
\arguments{
\item{nameVar}{name of the variables to add in the data frame}

\item{token}{a token from getToken function}

\item{smoothing}{logical, smoothing of the data}

\item{startDate}{}
}
\value{
data.frame
}
\description{
getDF
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken <- getToken("guest@opensilex.org","guest")
token <- aToken$data
getDF(token = token, nameVar = list("wind","temperature"))
}
}
