% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDF.R
\name{getDF}
\alias{getDF}
\title{getDF}
\usage{
getDF(nameVar, token, smoothing = FALSE, sensor = NULL,
  endDate = NULL, startDate = NULL)
}
\arguments{
\item{nameVar}{name of the variables to add in the data frame}

\item{token}{a token from \code{\link{getToken}} function}

\item{smoothing}{logical, smoothing of the data}

\item{sensor}{character, name of a sensor}

\item{endDate}{date, date from which to filter data, format "\%Y-\%m-\%dT\%H:\%M:\%S"}

\item{startDate}{date, date to which filter the data, format "\%Y-\%m-\%dT\%H:\%M:\%S"}
}
\value{
data.frame
}
\description{
get the dataset of incertitudes for the data selected
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
aToken <- getToken("guest@opensilex.org","guest")
token <- aToken$data
getDF(token = token, nameVar = list("wind","temperature"))
}
}
