% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_var_env.R
\name{getVarPretty}
\alias{getVarPretty}
\title{getVarPretty}
\usage{
getVarPretty(token)
}
\arguments{
\item{token}{(string) a token from getToken function}
}
\value{
WSResponse
}
\description{
getVarPretty
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 getVarPretty(token = token)
}
}
