% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarPretty.R
\name{getVarPretty}
\alias{getVarPretty}
\title{Get Variable's Names from WS2 and formate them}
\usage{
getVarPretty(token)
}
\arguments{
\item{token}{a token from \code{\link{getToken}} function}
}
\value{
WSResponse
}
\description{
Get Variable's Names from WS2 and formate them
}
\examples{
\donttest{
initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 getVarPretty(token = token)
}
}
