% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_var_env.R
\name{plotVar}
\alias{plotVar}
\title{plot.var}
\usage{
plotVar(nom_var, startDate = NULL, endDate = NULL, sensor = NULL,
  token)
}
\arguments{
\item{nom_var}{(string) name of the variable to plot}

\item{startDate}{(date format %Y-%m-%d) date from which to plot}

\item{endDate}{(date format %Y-%m-%d) date to which to plot}

\item{sensor}{(number) sensor's name that recorded the values}

\item{token}{(string) a token from getToken function}
}
\description{
plot.var
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "138.102.159.36:8080/uesAPI/rest/")
 aToken <- getToken("guest@phis.fr","guest")
 token <- aToken$data
 plotVar("temperature", token = aToken$data)
 plotVar("temperature", token = aToken$data, sensor = "s18002")
 plotVar("temperature", startDate = "2017-06-26", endDate = "2017-06-28", sensor = "s18002", token = aToken$data)
}
}
