% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu_var_env.R
\name{plotVar}
\alias{plotVar}
\title{plot.var}
\usage{
plotVar(nameVar, startDate = NULL, endDate = NULL, sensor = NULL,
  token)
}
\arguments{
\item{startDate}{(date format %Y-%m-%d) date from which to plot}

\item{endDate}{(date format %Y-%m-%d) date to which to plot}

\item{sensor}{(number) sensor's name that recorded the values}

\item{token}{(string) a token from getToken function}

\item{nomVar}{(list) name of the variable to plot}
}
\value{

}
\description{
plot.var
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 plotVar("temperature", token = token)
 plotVar("temperature", token = token, sensor = "s18002")
 plotVar("temperature", startDate = "2017-06-26", endDate = "2017-06-28", sensor = "s18002", token = token)
}

}
