% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarRel.R
\name{plotVarRel}
\alias{plotVarRel}
\title{Plot two Environmental variables relationship}
\usage{
plotVarRel(varX, varY, startDate = NULL, endDate = NULL, token,
  wsUrl = "www.opensilex.org/openSilexAPI/rest/")
}
\arguments{
\item{varX}{name of the variable to plot in X axis}

\item{varY}{name of the variable to plot in Y axis}

\item{startDate}{date from which to plot}

\item{endDate}{date to which to plot}

\item{token}{a token from \code{\link{getToken}} function}

\item{wsUrl}{url of the webservice}
}
\description{
Plot two Environmental variables relationship
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 plotVar("temperature", "radiation", token = token)
}

}
