% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarRel.R
\name{plotVarRel}
\alias{plotVarRel}
\title{Plot two Environmental variables relationship}
\usage{
plotVarRel(varX, varY, startDate = NULL, endDate = NULL,
  trend = FALSE, wsUrl = "www.opensilex.org/openSilexAPI/rest/", token)
}
\arguments{
\item{varX}{uri of the variable to plot in X axis, from the \code{\link{variableList}} function or the web service directly}

\item{varY}{uri of the variable to plot in Y axis, from the \code{\link{variableList}} function or the web service directly}

\item{startDate}{date from which to plot}

\item{endDate}{date to which to plot}

\item{trend}{logical, draw the trend of the scatterplot}

\item{wsUrl}{url of the webservice}

\item{token}{a token from \code{\link{getToken}} function}
}
\description{
Plot two Environmental variables relationship
}
\examples{
\donttest{
 initializeClientConnection(apiID="ws_private", url = "www.opensilex.org/openSilexAPI/rest/")
 aToken <- getToken("guest@opensilex.org","guest")
 token <- aToken$data
 vars <- variableList(token = token)
 vars
 plotVarRel( vars$value[1],
       vars$value[2],
       token = token,
       trend = TRUE)
}

}
